awallbackend/requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.2
django-cors-headers==4.7.0
django-csp==4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
packaging==24.2
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
